Implementing logginMiddleware to replace the logic in the Controllers' processLogin.
