+++ app.js
@@ -15,6 +15,7 @@ const cookieParser = require('cookie-parser');
 app.use(express.urlencoded({ extended: false}));
 app.use(express.json());
 app.use(session({ secret: 'Infousers' }));
+app.use(cookieParser());
 //app.use('/upload', cloudinaryMiddleware,)
