+++ middlewares/cookieMiddleware.js
@@ -1,0 +1,5 @@
+cookieMiddleware: (req, res, next) => {
+
+}
+
+module.exports = cookieMiddleware;
