+++ views/users/login.ejs
@@ -23,7 +23,7 @@
                         </div>
                         <div>
                             <div>
-                                <input type="checkbox" value="on">
+                                <input type="checkbox" value="on" name="saveUser">
                                 <label>Recordar</label>
                             </div>
                             <div>
