+++ views/users/userEditor.ejs
@@ -26,7 +26,7 @@
                         <input type="password" id="confirmPassword" name="confirmPassword" placeholder="Vuelva a ingresar su contraseña nueva" />
                     </div>
                 </div>
-                    <div class="items-registro"><label>Imagen producto:
+                    <div class="items-registro"><label>NUEVA IMAGEN DE PERFIL:
                         <input type="file" name="image" id="image"></label>
                     </div>
                 <div class="button-editor-product">
